requirements.txt
HTMLTestRunner-Python3==0.8.0
selenium==3.141.0
urllib3==1.26.4
